backend/requirements.txt
Flask>=2.1.0,<3
Werkzeug==2.2.3
SQLAlchemy==1.4.32
Flask-SQLAlchemy>=3.0.0
Flask-CORS==3.0.10
Flask-Migrate>=3.0.0
python-dotenv>=0.20.0
requests>=2.26.0
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.33.0
openai>=0.27.0
PyJWT>=2.3.0
